refactor(frontend): rename QRCode -> QrCode for ls-lint pascalcase
ls-lint requires strict PascalCase filenames; QRCode (consecutive capitals)
fails the .svelte pascalcase rule. Rename to QrCode, matching OISY's existing
convention (ReceiveQrCode, VipQrCodeModal, ...). Updates the 4 consumers.
