+++ src/frontend/src/lib/components/referral/ReferralCodeModal.svelte
@@ -10,7 +10,7 @@
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
 	import MessageBox from '$lib/components/ui/MessageBox.svelte';
-	import QRCode from '$lib/components/ui/QRCode.svelte';
+	import QrCode from '$lib/components/ui/QrCode.svelte';
 	import SkeletonQrCode from '$lib/components/ui/SkeletonQrCode.svelte';
 	import SkeletonText from '$lib/components/ui/SkeletonText.svelte';
 	import { OISY_REFERRAL_URL } from '$lib/constants/oisy.constants';
@@ -52,13 +52,13 @@
 	<ContentWithToolbar>
 		<div class="mx-auto mb-8 aspect-square h-80 max-h-[44vh] max-w-full rounded-xl bg-white p-4">
 			{#if nonNullish(referralCode)}
-				<QRCode value={referralUrl}>
+				<QrCode value={referralUrl}>
 					{#snippet logo()}
 						<div class="flex items-center justify-center rounded-lg bg-primary p-2">
 							<IconAstronautHelmet />
 						</div>
 					{/snippet}
-				</QRCode>
+				</QrCode>
 			{:else}
 				<SkeletonQrCode />
 			{/if}
