+++ src/frontend/src/lib/components/ui/QrCode.svelte
