+++ src/frontend/src/lib/components/vip/VipQrCodeModal.svelte
@@ -9,7 +9,7 @@
 	import ButtonCloseModal from '$lib/components/ui/ButtonCloseModal.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
-	import QRCode from '$lib/components/ui/QRCode.svelte';
+	import QrCode from '$lib/components/ui/QrCode.svelte';
 	import SkeletonText from '$lib/components/ui/SkeletonText.svelte';
 	import { CODE_REGENERATE_INTERVAL_IN_SECONDS } from '$lib/constants/app.constants';
 	import {
@@ -130,7 +130,7 @@
 	<ContentWithToolbar>
 		<div class="mx-auto mb-8 aspect-square h-80 max-h-[44vh] max-w-full rounded-xl bg-white p-4">
 			{#if nonNullish(code)}
-				<QRCode value={qrCodeUrl}>
+				<QrCode value={qrCodeUrl}>
 					{#snippet logo()}
 						<div class="flex items-center justify-center rounded-full bg-primary p-2">
 							{#if codeType === QrCodeType.VIP}
@@ -140,7 +140,7 @@
 							{/if}
 						</div>
 					{/snippet}
-				</QRCode>
+				</QrCode>
 			{/if}
 		</div>
